--- app/src/main/java/com/example/vinni_tracker/presentation/screens/home/HomeScreen.kt
@@ -1,378 +1,0 @@
-package com.example.vinni_tracker.presentation.screens.home
-
-import android.content.res.Configuration
-import androidx.compose.foundation.Image
-import androidx.compose.foundation.background
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.lazy.LazyRow
-import androidx.compose.foundation.lazy.items
-import androidx.compose.foundation.rememberScrollState
-import androidx.compose.foundation.verticalScroll
-import androidx.compose.material3.Card
-import androidx.compose.material3.CardDefaults
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Surface
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.painter.Painter
-import androidx.compose.ui.layout.ContentScale
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
-import androidx.lifecycle.viewmodel.compose.viewModel
-import androidx.navigation.NavHostController
-import androidx.navigation.compose.rememberNavController
-import com.example.vinni_tracker.R
-import com.example.vinni_tracker.data.HomeCardData
-import com.example.vinni_tracker.presentation.theme.VinniTrackerTheme
-
-// - reorganize the project structure to support desktop, tablet, smartwatch and mobile devices
-// - deal with the image extension
-// - add pull to refresh
-// - think about moving the login to another line in Title Block
-// or tell the user that there is a max number of characters (?)
-// - think about what to use background for the news
-// - reduce the distance between the navigation button and the recommendation button (?)
-@Composable
-fun HomeScreen(navController: NavHostController, modifier: Modifier = Modifier, viewModel: HomeViewModel = viewModel()) {
-  val greetingForTitle = stringResource(id = viewModel.greetingResId)
-  val backgroundForStudyTimer = painterResource(id = viewModel.studyTimerImageResId)
-  val cardDataStat = viewModel.cardDataStat
-  val shopNews = viewModel.shopNews
-
-  Column(
-    modifier = modifier
-      .fillMaxSize()
-      .background(MaterialTheme.colorScheme.background)
-      .padding(16.dp)
-      .verticalScroll(rememberScrollState()),
-    verticalArrangement = Arrangement.spacedBy(13.dp),
-  ) {
-    TitleBlock(greetingForTitle)
-    StudyTimerBlock(backgroundForStudyTimer, navController)
-    Statistic(cardDataStat, navController)
-    ShopNews(shopNews, navController)
-    StudyStorageBlock(navController)
-    DailyRecommendationBlock(navController)
-  }
-}
-
-@Composable
-fun TitleBlock(greeting: String, modifier: Modifier = Modifier) {
-  val login = "Ganesha"
-  val fullGreeting = stringResource(R.string.greeting_template, greeting, login) // pluralStringResource?
-  val wish = stringResource(R.string.wish_good_day)
-
-  Surface(
-    color = MaterialTheme.colorScheme.background,
-    modifier = modifier.fillMaxWidth(),
-  ) {
-    Row(
-      modifier = Modifier.padding(5.dp),
-      verticalAlignment = Alignment.CenterVertically,
-      horizontalArrangement = Arrangement.SpaceBetween,
-    ) {
-      Column {
-        Text(
-          text = fullGreeting,
-          style = MaterialTheme.typography.headlineSmall,
-          color = MaterialTheme.colorScheme.onBackground,
-        )
-        Text(
-          text = wish,
-          style = MaterialTheme.typography.bodyMedium,
-          color = MaterialTheme.colorScheme.onBackground,
-          modifier = Modifier.padding(top = 3.dp),
-        )
-      }
-      Image(
-        painter = painterResource(id = R.drawable.ic_cat),
-        contentDescription = "Cat Icon",
-        modifier = Modifier.size(50.dp),
-      )
-    }
-  }
-}
-
-@Composable
-fun StudyTimerBlock(image: Painter, navController: NavHostController, modifier: Modifier = Modifier) {
-  Card(
-    modifier = modifier
-      .fillMaxWidth()
-      .height(233.dp)
-      .clickable { navController.navigate("study_timer") },
-    shape = MaterialTheme.shapes.medium,
-    colors = CardDefaults.cardColors(
-      contentColor = MaterialTheme.colorScheme.onPrimaryContainer,
-    ),
-  ) {
-    Box(modifier = Modifier.fillMaxSize()) {
-      Image(
-        painter = image,
-        contentDescription = "Background image",
-        modifier = Modifier
-          .fillMaxSize(),
-        contentScale = ContentScale.Crop,
-      )
-      Column(
-        modifier = Modifier
-          .fillMaxSize()
-          .padding(16.dp),
-        horizontalAlignment = Alignment.CenterHorizontally,
-        verticalArrangement = Arrangement.Center,
-      ) {
-        Text(
-          text = "Study Timer",
-          style = MaterialTheme.typography.displayMedium,
-          color = MaterialTheme.colorScheme.onBackground,
-          textAlign = TextAlign.Center,
-        )
-      }
-    }
-  }
-}
-
-// fix the location
-@Composable
-fun Statistic(cardDataStat: List<HomeCardData>, navController: NavHostController, modifier: Modifier = Modifier) {
-  LazyRow(
-    modifier = modifier.fillMaxWidth(),
-    horizontalArrangement = Arrangement.spacedBy(16.dp),
-  ) {
-    items(cardDataStat) { card ->
-      Card(
-        modifier = Modifier
-          .width(179.dp)
-          .height(123.dp)
-          .clickable { navController.navigate(card.id) },
-        shape = MaterialTheme.shapes.medium,
-        colors = CardDefaults.cardColors(
-          containerColor = MaterialTheme.colorScheme.primaryContainer,
-          contentColor = MaterialTheme.colorScheme.onPrimaryContainer,
-        ),
-      ) {
-        Box(modifier = Modifier.fillMaxSize()) {
-          Image(
-            painter = painterResource(card.imageResId),
-            contentDescription = "Background image",
-            modifier = Modifier
-              .fillMaxSize(),
-            contentScale = ContentScale.Crop,
-          )
-          Column(
-            modifier = Modifier
-              .fillMaxSize()
-              .padding(8.dp),
-            horizontalAlignment = Alignment.CenterHorizontally,
-            verticalArrangement = Arrangement.Center,
-          ) {
-            Text(
-              text = card.title,
-              style = MaterialTheme.typography.titleMedium,
-              color = MaterialTheme.colorScheme.onPrimaryContainer,
-              textAlign = TextAlign.Center,
-            )
-            Text(
-              text = card.contentDescription,
-              style = MaterialTheme.typography.bodySmall,
-              color = MaterialTheme.colorScheme.onPrimaryContainer,
-              textAlign = TextAlign.Center,
-              modifier = Modifier.padding(top = 4.dp),
-            )
-          }
-        }
-      }
-    }
-  }
-}
-
-// fix the location
-// add background for news
-@Composable
-fun ShopNews(shopNews: List<HomeCardData>, navController: NavHostController, modifier: Modifier = Modifier) {
-  LazyRow(
-    modifier = modifier.fillMaxWidth(),
-    horizontalArrangement = Arrangement.spacedBy(16.dp),
-  ) {
-    items(shopNews) { card ->
-      Card(
-        modifier = Modifier
-          .width(179.dp)
-          .height(123.dp)
-          .clickable { navController.navigate(card.id) },
-        shape = MaterialTheme.shapes.medium,
-        colors = CardDefaults.cardColors(
-          containerColor = MaterialTheme.colorScheme.primaryContainer,
-          contentColor = MaterialTheme.colorScheme.onPrimaryContainer,
-        ),
-      ) {
-        Box(modifier = Modifier.fillMaxSize()) {
-          when (card.id) {
-            "shop_screen" -> {
-              Text(
-                text = " HOT\nPRICE ",
-                style = MaterialTheme.typography.labelSmall.copy(
-                  color = Color(0xFFF3BE3B),
-                  fontWeight = FontWeight.Bold,
-                  fontSize = 10.sp,
-                ),
-                modifier = Modifier
-                  .align(Alignment.TopStart)
-                  .padding(8.dp),
-              )
-              Column(
-                modifier = Modifier.align(Alignment.Center),
-                horizontalAlignment = Alignment.CenterHorizontally,
-                verticalArrangement = Arrangement.Center,
-              ) {
-                Image(
-                  painter = painterResource(id = card.imageResId),
-                  contentDescription = "${card.title} icon",
-                  modifier = Modifier.size(48.dp),
-                  contentScale = ContentScale.Fit,
-                )
-                Text(
-                  text = card.title,
-                  style = MaterialTheme.typography.bodyMedium.copy(fontWeight = FontWeight.Medium),
-                  color = MaterialTheme.colorScheme.onPrimaryContainer,
-                  textAlign = TextAlign.Center,
-                )
-              }
-            }
-            "news_screen" -> {
-              Card(
-                modifier = Modifier
-                  .fillMaxWidth()
-                  .height(95.dp)
-                  .clickable { navController.navigate("news_screen") },
-                shape = MaterialTheme.shapes.medium,
-                colors = CardDefaults.cardColors(
-                  containerColor = MaterialTheme.colorScheme.primaryContainer,
-                  contentColor = MaterialTheme.colorScheme.onPrimaryContainer,
-                ),
-              ) {
-                Box(modifier = Modifier.fillMaxSize()) {
-                  Column(
-                    modifier = Modifier.align(Alignment.Center),
-                    horizontalAlignment = Alignment.CenterHorizontally,
-                    verticalArrangement = Arrangement.Center,
-                  ) {
-                    Text(
-                      text = "News",
-                      style = MaterialTheme.typography.titleMedium,
-                      color = MaterialTheme.colorScheme.onPrimaryContainer,
-                      modifier = Modifier.padding(top = 20.dp),
-                    )
-                    Text(
-                      text = "Latest updates",
-                      style = MaterialTheme.typography.bodySmall,
-                      color = MaterialTheme.colorScheme.onPrimaryContainer,
-                      modifier = Modifier.padding(top = 4.dp),
-                    )
-                  }
-                }
-              }
-            }
-          }
-        }
-      }
-    }
-  }
-}
-
-@Composable
-fun StudyStorageBlock(navController: NavHostController, modifier: Modifier = Modifier) {
-  Card(
-    modifier = modifier
-      .fillMaxWidth()
-      .height(45.dp)
-      .clickable { navController.navigate("study_storage") },
-    shape = MaterialTheme.shapes.medium,
-    colors = CardDefaults.cardColors(
-      containerColor = MaterialTheme.colorScheme.primaryContainer,
-      contentColor = MaterialTheme.colorScheme.onPrimaryContainer,
-    ),
-  ) {
-    Column(
-      modifier = Modifier
-        .fillMaxSize()
-        .padding(10.dp),
-      horizontalAlignment = Alignment.Start,
-    ) {
-      Text(
-        text = "Study storage",
-        style = MaterialTheme.typography.titleMedium,
-        color = MaterialTheme.colorScheme.onPrimaryContainer,
-      )
-    }
-  }
-}
-
-@Composable
-fun DailyRecommendationBlock(navController: NavHostController, modifier: Modifier = Modifier) {
-  Card(
-    modifier = modifier
-      .fillMaxWidth()
-      .height(95.dp)
-      .clickable { navController.navigate("recommendation_screen") },
-    shape = MaterialTheme.shapes.medium,
-    colors = CardDefaults.cardColors(
-      containerColor = MaterialTheme.colorScheme.primaryContainer,
-      contentColor = MaterialTheme.colorScheme.onPrimaryContainer,
-    ),
-  ) {
-    Box(modifier = Modifier.fillMaxSize()) {
-      Image(
-        painter = painterResource(id = R.drawable.daily_news_afternoon),
-        contentDescription = "Background image",
-        modifier = Modifier
-          .fillMaxSize(),
-        contentScale = ContentScale.Crop,
-      )
-      Column(
-        modifier = Modifier
-          .fillMaxSize()
-          .padding(16.dp),
-        horizontalAlignment = Alignment.Start,
-      ) {
-        Text(
-          text = "Daily recommendation",
-          style = MaterialTheme.typography.titleMedium,
-          color = MaterialTheme.colorScheme.onPrimaryContainer,
-        )
-        Text(
-          text = "Meditation: 3-10 min",
-          style = MaterialTheme.typography.bodySmall,
-          color = MaterialTheme.colorScheme.onPrimaryContainer,
-          modifier = Modifier.padding(top = 4.dp),
-        )
-      }
-    }
-  }
-}
-
-// locale = "fr-rFR" - deal with translation
-@Preview(showBackground = true, uiMode = Configuration.UI_MODE_NIGHT_NO)
-@Composable
-private fun HomeScreenPreview() {
-  VinniTrackerTheme(dynamicColor = false, darkTheme = false) {
-    HomeScreen(navController = rememberNavController())
-  }
-}
