+++ app/src/main/java/com/example/vinni_tracker/ui/navigation/BottomNavItem.kt
@@ -1,0 +1,10 @@
+package com.example.vinni_tracker.ui.navigation
+
+data class BottomNavItem(
+  // Text below icon
+  val label: String,
+  // Icon
+  val icon: Int,
+  // Route to the specific screen
+  val route: String,
+)
