+++ app/src/main/java/com/example/vinni_tracker/ui/navigation/BottomNavigationBar.kt
@@ -1,0 +1,280 @@
+package com.example.vinni_tracker.ui.navigation
+
+import android.content.res.Configuration
+import androidx.compose.animation.AnimatedContent
+import androidx.compose.animation.core.Spring
+import androidx.compose.animation.core.animateFloat
+import androidx.compose.animation.core.animateFloatAsState
+import androidx.compose.animation.core.animateIntOffset
+import androidx.compose.animation.core.snap
+import androidx.compose.animation.core.spring
+import androidx.compose.animation.core.updateTransition
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.offset
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.material3.Icon
+import androidx.compose.material3.NavigationBar
+import androidx.compose.material3.NavigationBarItem
+import androidx.compose.material3.NavigationBarItemDefaults
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableIntStateOf
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.saveable.rememberSaveable
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.alpha
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.geometry.Rect
+import androidx.compose.ui.geometry.Size
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.Outline
+import androidx.compose.ui.graphics.Path
+import androidx.compose.ui.graphics.Shape
+import androidx.compose.ui.graphics.graphicsLayer
+import androidx.compose.ui.layout.onPlaced
+import androidx.compose.ui.platform.LocalDensity
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.Density
+import androidx.compose.ui.unit.Dp
+import androidx.compose.ui.unit.IntOffset
+import androidx.compose.ui.unit.IntSize
+import androidx.compose.ui.unit.LayoutDirection
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.zIndex
+import androidx.navigation.NavHostController
+import androidx.navigation.compose.currentBackStackEntryAsState
+import androidx.navigation.compose.rememberNavController
+import com.example.vinni_tracker.ui.theme.VinniTrackerTheme
+
+// take the colors out of here
+// deal with the animation of the circle (make it faster)
+@Composable
+fun BottomNavigationBar(navController: NavHostController, modifier: Modifier = Modifier) {
+  val circleRadius = 25.dp
+  var selectedItem by rememberSaveable { mutableIntStateOf(0) }
+  var barSize by remember { mutableStateOf(IntSize(0, 0)) }
+  val offsetStep = remember(barSize) {
+    barSize.width.toFloat() / (Constants.BottomNavItems.size * 2)
+  }
+  val offset = remember(selectedItem, offsetStep) {
+    offsetStep + selectedItem * 2 * offsetStep
+  }
+  val circleRadiusPx = LocalDensity.current.run { circleRadius.toPx().toInt() }
+  val offsetTransition = updateTransition(offset, "offset transition")
+  val animation = spring<Float>(dampingRatio = 0.5f, stiffness = Spring.StiffnessVeryLow)
+  val cutoutOffset by offsetTransition.animateFloat(
+    transitionSpec = {
+      if (this.initialState == 0f) snap() else animation
+    },
+    label = "cutout offset",
+  ) { it }
+  val pxValue = with(LocalDensity.current) { 10.dp.toPx().toInt() }
+  val circleOffset by offsetTransition.animateIntOffset(
+    transitionSpec = {
+      if (this.initialState == 0f) snap() else spring(animation.dampingRatio, animation.stiffness)
+    },
+    label = "circle offset",
+  ) {
+    IntOffset(it.toInt() - circleRadiusPx, -circleRadiusPx + pxValue)
+  }
+  val barShape = remember(cutoutOffset) {
+    BarShape(
+      offset = cutoutOffset,
+      circleRadius = circleRadius,
+      cornerRadius = 0.dp, // here can round the corners of the bar (for some reason it doesn't work now)
+    )
+  }
+
+  Box(modifier = Modifier.fillMaxWidth()) {
+    NavigationBar(
+      containerColor = Color(0xFF166083),
+      modifier = Modifier.height(90.dp),
+    ) {
+      val navBackStackEntry by navController.currentBackStackEntryAsState()
+      val currentRoute = navBackStackEntry?.destination?.route
+
+      Row(
+        modifier = Modifier
+          .onPlaced { barSize = it.size }
+          .graphicsLayer {
+            shape = barShape
+            clip = true
+          }
+          .fillMaxWidth()
+          .background(Color(0xFF166083)),
+        horizontalArrangement = Arrangement.SpaceAround,
+      ) {
+        Constants.BottomNavItems.forEachIndexed { index, navItem ->
+          val isSelected = currentRoute == navItem.route
+          NavigationBarItem(
+            selected = isSelected,
+            onClick = {
+              selectedItem = index
+              navController.navigate(navItem.route) {
+                // Avoid multiple copies of the same destination when
+                // reselecting the same item
+                launchSingleTop = true
+                // Restore state when reselecting a previously selected item
+                restoreState = true
+              }
+            },
+            icon = {
+              val iconAlpha by animateFloatAsState(
+                targetValue = if (isSelected) 0f else 1f,
+                label = "Navbar item icon",
+              )
+              Icon(
+                painter = painterResource(id = navItem.icon),
+                contentDescription = navItem.label,
+                modifier = Modifier.alpha(iconAlpha),
+              )
+            },
+            label = { Text(text = navItem.label) },
+            alwaysShowLabel = true,
+            colors = NavigationBarItemDefaults.colors(
+              selectedIconColor = Color(0xFFD1F0FF),
+              selectedTextColor = Color(0xFFD1F0FF),
+              unselectedIconColor = Color(0xFFFFE7BF),
+              unselectedTextColor = Color(0xFFFFE7BF),
+              indicatorColor = Color.Transparent,
+            ),
+          )
+        }
+      }
+    }
+    Circle(
+      modifier = Modifier
+        .offset { circleOffset }
+        .zIndex(1f),
+      color = Color(0xFF026D9E),
+      radius = circleRadius,
+      button = Constants.BottomNavItems[selectedItem],
+      iconColor = Color(0xFFD1F0FF),
+    )
+  }
+}
+
+private class BarShape(
+  private val offset: Float,
+  private val circleRadius: Dp,
+  private val cornerRadius: Dp,
+  private val circleGap: Dp = 5.dp,
+) : Shape {
+  override fun createOutline(size: Size, layoutDirection: LayoutDirection, density: Density): Outline {
+    return Outline.Generic(getPath(size, density))
+  }
+  private fun getPath(size: Size, density: Density): Path {
+    val cutoutCenterX = offset
+    val cutoutRadius = density.run { (circleRadius + circleGap).toPx() }
+    val cornerRadiusPx = density.run { cornerRadius.toPx() }
+    val cornerDiameter = cornerRadiusPx * 2
+    return Path().apply {
+      val cutoutEdgeOffset = cutoutRadius * 1.5f
+      val cutoutLeftX = cutoutCenterX - cutoutEdgeOffset
+      val cutoutRightX = cutoutCenterX + cutoutEdgeOffset
+
+      // Bottom left
+      moveTo(x = 0f, y = size.height)
+      // Top left
+      if (cutoutLeftX > 0) {
+        val realLeftCornerDiameter = if (cutoutLeftX >= cornerRadiusPx) {
+          cornerDiameter
+        } else {
+          cutoutLeftX * 2
+        }
+        arcTo(
+          rect = Rect(
+            left = 0f,
+            top = 0f,
+            right = realLeftCornerDiameter,
+            bottom = realLeftCornerDiameter,
+          ),
+          startAngleDegrees = 180.0f,
+          sweepAngleDegrees = 90.0f,
+          forceMoveTo = false,
+        )
+      }
+      lineTo(cutoutLeftX, 0f)
+      // Cutout
+      cubicTo(
+        x1 = cutoutCenterX - cutoutRadius,
+        y1 = 0f,
+        x2 = cutoutCenterX - cutoutRadius,
+        y2 = cutoutRadius,
+        x3 = cutoutCenterX,
+        y3 = cutoutRadius,
+      )
+      cubicTo(
+        x1 = cutoutCenterX + cutoutRadius,
+        y1 = cutoutRadius,
+        x2 = cutoutCenterX + cutoutRadius,
+        y2 = 0f,
+        x3 = cutoutRightX,
+        y3 = 0f,
+      )
+      // Top right
+      if (cutoutRightX < size.width) {
+        val realRightCornerDiameter = if (cutoutRightX <= size.width - cornerRadiusPx) {
+          cornerDiameter
+        } else {
+          (size.width - cutoutRightX) * 2
+        }
+        arcTo(
+          rect = Rect(
+            left = size.width - realRightCornerDiameter,
+            top = 0f,
+            right = size.width,
+            bottom = realRightCornerDiameter,
+          ),
+          startAngleDegrees = -90.0f,
+          sweepAngleDegrees = 90.0f,
+          forceMoveTo = false,
+        )
+      }
+      // Bottom right
+      lineTo(x = size.width, y = size.height)
+      close()
+    }
+  }
+}
+
+@Composable
+private fun Circle(radius: Dp, button: BottomNavItem, iconColor: Color, modifier: Modifier = Modifier, color: Color = Color.White) {
+  Box(
+    contentAlignment = Alignment.Center,
+    modifier = modifier
+      .size(radius * 2)
+      .clip(CircleShape)
+      .background(color),
+  ) {
+    AnimatedContent(
+      targetState = button.icon,
+      label = "Bottom bar circle icon",
+    ) { targetIcon ->
+      Icon(
+        painter = painterResource(id = targetIcon),
+        contentDescription = button.label,
+        tint = iconColor,
+      )
+    }
+  }
+}
+
+@Preview(showBackground = true, uiMode = Configuration.UI_MODE_NIGHT_NO)
+@Composable
+private fun BottomNavigationBarPreview() {
+  VinniTrackerTheme(dynamicColor = false, darkTheme = false) {
+    BottomNavigationBar(navController = rememberNavController())
+  }
+}
