+++ app/src/main/java/com/example/vinni_tracker/ui/screens/home/components/DailyRecommendationSection.kt
@@ -1,0 +1,65 @@
+package com.example.vinni_tracker.ui.screens.home.components
+
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.Card
+import androidx.compose.material3.CardDefaults
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.unit.dp
+import androidx.navigation.NavHostController
+import com.example.vinni_tracker.R
+
+@Composable
+fun DailyRecommendationSection(navController: NavHostController, modifier: Modifier = Modifier) {
+  Card(
+    modifier = modifier
+      .fillMaxWidth()
+      .height(95.dp)
+      .clickable { navController.navigate("recommendation_screen") },
+    shape = MaterialTheme.shapes.medium,
+    colors = CardDefaults.cardColors(
+      containerColor = MaterialTheme.colorScheme.primaryContainer,
+      contentColor = MaterialTheme.colorScheme.onPrimaryContainer,
+    ),
+  ) {
+    Box(modifier = Modifier.fillMaxSize()) {
+      Image(
+        painter = painterResource(id = R.drawable.daily_news_afternoon),
+        contentDescription = "Background image",
+        modifier = Modifier
+          .fillMaxSize(),
+        contentScale = ContentScale.Crop,
+      )
+      Column(
+        modifier = Modifier
+          .fillMaxSize()
+          .padding(16.dp),
+        horizontalAlignment = Alignment.Start,
+      ) {
+        Text(
+          text = "Daily recommendation",
+          style = MaterialTheme.typography.titleMedium,
+          color = MaterialTheme.colorScheme.onPrimaryContainer,
+        )
+        Text(
+          text = "Meditation: 3-10 min",
+          style = MaterialTheme.typography.bodySmall,
+          color = MaterialTheme.colorScheme.onPrimaryContainer,
+          modifier = Modifier.padding(top = 4.dp),
+        )
+      }
+    }
+  }
+}
