+++ app/src/main/java/com/example/vinni_tracker/ui/screens/home/components/StudyStorageSection.kt
@@ -1,0 +1,45 @@
+package com.example.vinni_tracker.ui.screens.home.components
+
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.Card
+import androidx.compose.material3.CardDefaults
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+import androidx.navigation.NavHostController
+
+@Composable
+fun StudyStorageSection(navController: NavHostController, modifier: Modifier = Modifier) {
+  Card(
+    modifier = modifier
+      .fillMaxWidth()
+      .height(45.dp)
+      .clickable { navController.navigate("study_storage") },
+    shape = MaterialTheme.shapes.medium,
+    colors = CardDefaults.cardColors(
+      containerColor = MaterialTheme.colorScheme.primaryContainer,
+      contentColor = MaterialTheme.colorScheme.onPrimaryContainer,
+    ),
+  ) {
+    Column(
+      modifier = Modifier
+        .fillMaxSize()
+        .padding(10.dp),
+      horizontalAlignment = Alignment.Start,
+    ) {
+      Text(
+        text = "Study storage",
+        style = MaterialTheme.typography.titleMedium,
+        color = MaterialTheme.colorScheme.onPrimaryContainer,
+      )
+    }
+  }
+}
