+++ app/src/main/java/com/example/vinni_tracker/ui/screens/home/storage/StudyStorageScreen.kt
@@ -1,0 +1,33 @@
+package com.example.vinni_tracker.ui.screens.home.storage
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Person
+import androidx.compose.material3.Icon
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+
+@Composable
+fun StudyStorageScreen(modifier: Modifier = Modifier) {
+  Column(
+    modifier = Modifier
+      .fillMaxSize()
+      .background(Color.White),
+    horizontalAlignment = Alignment.CenterHorizontally,
+    verticalArrangement = Arrangement.Center,
+  ) {
+    Icon(
+      imageVector = Icons.Default.Person,
+      contentDescription = "Profile",
+      tint = Color(0xFF0F9D58),
+    )
+
+    Text(text = "Profile", color = Color.Black)
+  }
+}
