+++ app/src/main/java/com/example/vinni_tracker/ui/theme/Color.kt
@@ -1,0 +1,35 @@
+package com.example.vinni_tracker.ui.theme
+
+import androidx.compose.ui.graphics.Color
+
+// LIGHT THEME
+val PrimaryLight = Color(0xFFF44336) // main color
+val SecondaryLight = Color(0xFF9C27B0)
+val TertiaryLight = Color(0xFF673AB7)
+val OnPrimaryLight = Color(0xFF3F51B5)
+val OnSecondaryLight = Color(0xFF2196F3)
+val OnTertiaryLight = Color(0xFF010D10)
+val BackgroundLight = Color(0xFFE2EFF2)
+val SurfaceLight = Color(0xFF4CAF50)
+val OnBackgroundLight = Color(0xFF32343E)
+val OnSurfaceLight = Color(0xFFFFEB3B)
+val PrimaryContainerLight = Color(0xFF049CE1)
+val OnPrimaryContainerLight = Color(0xFFFFE7BF)
+val SecondaryContainerLight = Color(0xFFFFE7BF)
+val OnSecondaryContainerLight = Color(0xFF3F414E)
+
+// DARK THEME
+val PrimaryDark = Color(0xFFD1F0FF)
+val SecondaryDark = Color(0xFF024A6B)
+val TertiaryDark = Color(0xFFA5CDEA)
+val OnPrimaryDark = Color(0xFF010D10)
+val OnSecondaryDark = Color(0xFF010D10)
+val OnTertiaryDark = Color(0xFF010D10)
+val BackgroundDark = Color(0xFF1C1B1F)
+val SurfaceDark = Color(0xFF1C1B1F)
+val OnBackgroundDark = Color(0xFFE6E1E5)
+val OnSurfaceDark = Color(0xFFE6E1E5)
+val PrimaryContainerDark = Color(0xFFFFEB3B)
+val OnPrimaryContainerDark = Color(0xFFFFE7BF)
+val SecondaryContainerDark = Color(0xFFFFE7BF)
+val OnSecondaryContainerDark = Color(0xFF3F414E)
